# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/support/ping": {
            "get": {
                "description": "Check if API is up returning \"OK\"",
                "tags": [
                    "Support"
                ],
                "responses": {}
            }
        },
        "/user/add": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a new User",
                "parameters": [
                    {
                        "description": "Add user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.ResponseBody"
                        }
                    }
                }
            }
        },
        "/user/del/{id}": {
            "delete": {
                "description": "Delete an user by its id",
                "tags": [
                    "User"
                ],
                "summary": "Delete an user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID User",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.ResponseBody"
                        }
                    }
                }
            }
        },
        "/user/find": {
            "get": {
                "description": "Return a user by its id, name or age as query params",
                "tags": [
                    "User"
                ],
                "summary": "Return a user by id, name or age",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search by id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by age",
                        "name": "age",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.ResponseBody"
                        }
                    }
                }
            }
        },
        "/user/list": {
            "get": {
                "description": "Return a list of user objects",
                "tags": [
                    "User"
                ],
                "summary": "Return all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.ResponseBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "routes.ResponseBody": {
            "type": "object",
            "properties": {
                "body": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "routes.User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}